Refactor ProtoFactory
