+++ src/index.ts
@@ -7,11 +7,11 @@ import Interface from "./commands/Interface";
 import HttpServer from "./http/HttpServer";
 import SRServer from "./server/kcp/SRServer";
 import Logger from "./util/Logger";
-import ProtoFactory from "./ProtoFactory"
+import ProtoFactory from "./util/ProtoFactory"
 
 const c = new Logger("CrepeSR");
-ProtoFactory.init();
 c.log(`Starting CrepeSR...`);
+ProtoFactory.init();
 Interface.start();
 HttpServer.getInstance().start();
 SRServer.getInstance().start();
